Add Rust bench
Crazy stuff: if you compile without `-O`, the bench takes *2 minutes*. With
`-O` it drops to `12s` for me. Pretty scary / impressive!

Also, Rust looks funny sometimes, though it's probably me doing somethign
in a very bad way:

        let v = std::str::from_utf8(cur_text.as_slice()).unwrap().to_owned();

The pointers are tedious to mangle, but worth the safety. All the reported
compiler warnings / errors are also pretty sweet. If they don't bloat the lang
it will be great when it matures :-)